refacto(tripFeatureCard): add type & add full type for maps
