+++ move_together_app/lib/Widgets/Card/trip_feature_card_empty_body.dart
@@ -1,0 +1,27 @@
+import 'package:flutter/material.dart';
+
+class TripFeatureCardEmptyBody extends StatelessWidget {
+  final String message;
+
+  const TripFeatureCardEmptyBody({
+    super.key, 
+    required this.message,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Center(
+      child: Padding(
+        padding: const EdgeInsets.symmetric(horizontal: 48),
+        child: Text(
+          message,
+          style: const TextStyle(
+            color: Colors.black54,
+            fontSize: 12,
+          ),
+          textAlign: TextAlign.center,
+        ),
+      ),
+    );
+  }
+}
